+++ doc/source/changelog.rst
@@ -5,6 +5,8 @@ v0.0.3
 ^^^^^^
 
 - Pages can be searched from the right top search box.
+- :rst:dir:`table-data-and-image` highlights data with different
+  values.
 - :rst:dir:`list-pages` shows sub-pages as a nested list, not as a
   flat list.
 - Show sub-pages when the accessed page does not exist but at least